Precursor spawning
Spawn points for enemy and player units placed. Tags for units and bases made.
The idea for the script will be to have the units spawn in, search for objects tagged as 'EnemyBase' and 'PlayerBase' respectively and then path toward them. Collision will trigger checking the tag of what was hit and proceeding from there.